+++ packages/brightcove-video/__tests__/brightcove-video.test.js
@@ -14,19 +14,23 @@ const playIconEmoji = () => (
   </span>
 );
 
+const defaultVideoProps = {
+  policyKey,
+  videoId,
+  accountId,
+  poster: {
+    uri:
+      "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
+  }
+};
+
+beforeEach(() => {
+  BrightcoveVideo.activePlayers = [];
+});
+
 it("renders poster correctly before launch", () => {
   const tree = renderer
-    .create(
-      <BrightcoveVideo
-        policyKey={policyKey}
-        videoId={videoId}
-        accountId={accountId}
-        poster={{
-          uri:
-            "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-        }}
-      />
-    )
+    .create(<BrightcoveVideo {...defaultVideoProps} />)
     .toJSON();
 
   expect(tree).toMatchSnapshot();
@@ -35,16 +39,7 @@ it("renders poster correctly before launch", () => {
 it("renders poster with custom play icon if specified", () => {
   const tree = renderer
     .create(
-      <BrightcoveVideo
-        policyKey={policyKey}
-        videoId={videoId}
-        accountId={accountId}
-        playIcon={playIconEmoji()}
-        poster={{
-          uri:
-            "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-        }}
-      />
+      <BrightcoveVideo {...defaultVideoProps} playIcon={playIconEmoji()} />
     )
     .toJSON();
 
@@ -52,17 +47,7 @@ it("renders poster with custom play icon if specified", () => {
 });
 
 it("will launch if play is called", () => {
-  const root = renderer.create(
-    <BrightcoveVideo
-      policyKey={policyKey}
-      videoId={videoId}
-      accountId={accountId}
-      poster={{
-        uri:
-          "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-      }}
-    />
-  );
+  const root = renderer.create(<BrightcoveVideo {...defaultVideoProps} />);
 
   const rootInstance = root.getInstance();
 
@@ -71,18 +56,34 @@ it("will launch if play is called", () => {
   expect(root.toJSON()).toMatchSnapshot();
 });
 
+it("pauses other playing videos if play is called", () => {
+  renderer.create(<BrightcoveVideo {...defaultVideoProps} />);
+  renderer.create(<BrightcoveVideo {...defaultVideoProps} />);
+
+  const [component1, component2] = BrightcoveVideo.activePlayers;
+  jest.spyOn(component1, "pause");
+  jest.spyOn(component2, "pause");
+
+  component1.handlePlay();
+  expect(component1.pause).not.toHaveBeenCalled();
+  component2.handlePlay();
+  expect(component1.pause).toHaveBeenCalled();
+});
+
+it("doesn't hold references to players after they have been unmounted", () => {
+  renderer.create(<BrightcoveVideo {...defaultVideoProps} />);
+  const p2 = renderer.create(<BrightcoveVideo {...defaultVideoProps} />);
+
+  expect(BrightcoveVideo.activePlayers.length).toBe(2);
+  const [component1] = BrightcoveVideo.activePlayers;
+  p2.unmount();
+  expect(BrightcoveVideo.activePlayers.length).toBe(1);
+  expect(BrightcoveVideo.activePlayers[0]).toBe(component1);
+});
+
 it("will reset properly", () => {
   const root = renderer.create(
-    <BrightcoveVideo
-      policyKey={policyKey}
-      videoId={videoId}
-      accountId={accountId}
-      poster={{
-        uri:
-          "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-      }}
-      autoplay
-    />
+    <BrightcoveVideo {...defaultVideoProps} autoplay />
   );
 
   const rootInstance = root.getInstance();
@@ -94,16 +95,7 @@ it("will reset properly", () => {
 
 it("will reset if 'resetOnFinish' is true & video finishes", done => {
   const root = renderer.create(
-    <BrightcoveVideo
-      policyKey={policyKey}
-      videoId={videoId}
-      accountId={accountId}
-      poster={{
-        uri:
-          "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-      }}
-      resetOnFinish
-    />
+    <BrightcoveVideo {...defaultVideoProps} resetOnFinish />
   );
 
   const rootInstance = root.getInstance();
@@ -121,16 +113,7 @@ it("will reset if 'resetOnFinish' is true & video finishes", done => {
 
 it("will call child components play and pause methods if child component is ready", () => {
   const root = renderer.create(
-    <BrightcoveVideo
-      policyKey={policyKey}
-      videoId={videoId}
-      accountId={accountId}
-      poster={{
-        uri:
-          "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-      }}
-      autoplay
-    />
+    <BrightcoveVideo {...defaultVideoProps} autoplay />
   );
 
   const rootInstance = root.getInstance();
@@ -153,16 +136,7 @@ it("will call child components play and pause methods if child component is read
 
 it("will call native fullscreen player if 'directToFullscreen option passed'", () => {
   const root = renderer.create(
-    <BrightcoveVideo
-      policyKey={policyKey}
-      videoId={videoId}
-      accountId={accountId}
-      poster={{
-        uri:
-          "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-      }}
-      directToFullscreen
-    />
+    <BrightcoveVideo {...defaultVideoProps} directToFullscreen />
   );
 
   const rootInstance = root.getInstance();
@@ -180,16 +154,7 @@ it("will call native fullscreen player if 'directToFullscreen option passed'", (
 
 it("will handle an error properly", () => {
   const root = renderer.create(
-    <BrightcoveVideo
-      policyKey={policyKey}
-      videoId={videoId}
-      accountId={accountId}
-      poster={{
-        uri:
-          "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-      }}
-      autoplay
-    />
+    <BrightcoveVideo {...defaultVideoProps} autoplay />
   );
 
   const rootInstance = root.getInstance();
@@ -201,16 +166,7 @@ it("will handle an error properly", () => {
 
 it("the player can trigger errors", () => {
   const root = renderer.create(
-    <BrightcoveVideo
-      policyKey={policyKey}
-      videoId={videoId}
-      accountId={accountId}
-      poster={{
-        uri:
-          "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-      }}
-      autoplay
-    />
+    <BrightcoveVideo {...defaultVideoProps} autoplay />
   );
 
   const rootInstance = root.getInstance();