+++ packages/link/src/link.web.js
@@ -8,7 +8,14 @@ const RespLink = responsiveLinkStyles =>
     mediumUp: () => `${responsiveLinkStyles.medium}`
   });
 
-const Link = ({ index, url, onPress, children, responsiveLinkStyles }) => {
+const Link = ({
+  index,
+  url,
+  onPress,
+  children,
+  responsiveLinkStyles,
+  target
+}) => {
   const Wrapper =
     responsiveLinkStyles !== null ? RespLink(responsiveLinkStyles) : "a";
 
@@ -17,10 +24,19 @@ const Link = ({ index, url, onPress, children, responsiveLinkStyles }) => {
       ? { textDecoration: "underline" }
       : { textDecoration: "none" };
 
-  return (
-    <Wrapper key={index} href={url} onClick={onPress} style={style}>
+  const props = {
+    key: index,
+    href: url,
+    onClick: onPress,
+    style
+  };
+
+  return target ? (
+    <Wrapper {...props} target={target}>
       {children}
     </Wrapper>
+  ) : (
+    <Wrapper {...props}>{children}</Wrapper>
   );
 };
 
@@ -32,12 +48,14 @@ Link.propTypes = {
     base: PropTypes.string,
     medium: PropTypes.string
   }),
-  index: PropTypes.string
+  index: PropTypes.string,
+  target: PropTypes.string
 };
 
 Link.defaultProps = {
   responsiveLinkStyles: null,
-  index: "0"
+  index: "0",
+  target: null
 };
 
 export default Link;