+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.9"></a>
+## [0.0.9](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.8...@times-components/related-articles@0.0.9) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.8"></a>
+## [0.0.8](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.7...@times-components/related-articles@0.0.8) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.7"></a>
+## [0.0.7](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.5...@times-components/related-articles@0.0.7) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.6"></a>
+## [0.0.6](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.5...@times-components/related-articles@0.0.6) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.0.5"></a>
+## [0.0.5](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.4...@times-components/related-articles@0.0.5) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
 <a name="0.0.4"></a>
 ## [0.0.4](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.0.3...@times-components/related-articles@0.0.4) (2018-04-06)